+++ src/readoutConfigEditor.tcl
@@ -152,6 +152,8 @@ set configurationParametersDescriptor {
 | readout | disableAggregatorSlicing | int | 0 | When set, the aggregator slicing is disabled, data pages are passed through without grouping/slicing. |
 | readout | disableTimeframes | int | 0 | When set, all timeframe related features are disabled (this may supersede other config parameters). |
 | readout | exitTimeout | double | -1 | Time in seconds after which the program exits automatically. -1 for unlimited. |
+| readout | externalSyncServer | string | | If set, ZMQ address to request SYNC signal at SOR. |
+| readout | externalSyncTimeout | int | 3000 | Timeout (in milliseconds) to wait for the SYNC signal at SOR (when externalSyncServer is defined). |
 | readout | fairmqConsoleSeverity | int | -1 | Select amount of FMQ messages with fair::Logger::SetConsoleSeverity(). Value as defined in Severity enum defined from FairLogger/Logger.h. Use -1 to leave current setting. |
 | readout | flushConsumerTimeout | double | 1 | Time in seconds to wait before stopping the consumers (ie wait allocated pages released). 0 means stop immediately. |
 | readout | flushEquipmentTimeout | double | 1 | Time in seconds to wait for data once the equipments are stopped. 0 means stop immediately. |