+++ src/testMonitor.cxx
@@ -21,18 +21,20 @@ int main()
 {
 
   std::unique_ptr<Monitoring> monitoringCollector;
-  monitoringCollector = MonitoringFactory::Get("influxdb-udp://aido2mon-gpn.cern.ch:8088");
-  monitoringCollector->enableProcessMonitoring(1);
+  //monitoringCollector = MonitoringFactory::Get("influxdb-udp://aido2mon-gpn.cern.ch:8088");
+  //monitoringCollector = MonitoringFactory::Get("influxdb-unix:///tmp/unix_dgram_server.sock");
+  monitoringCollector = MonitoringFactory::Get("influxdb-stdout:///tmp/telegraf.sock");
+  //monitoringCollector->enableProcessMonitoring(1);
 
   uint64_t bytesTotal = 0;
 
-  for (;;) {
-    bytesTotal += 1000000000;
+  for (int n=0;n<100000;n++) {
+    bytesTotal += 1;
     monitoringCollector->send({ bytesTotal, "readout.BytesTotal" });
-    monitoringCollector->send({ bytesTotal, "readout.BytesTotal" }, DerivedMetricMode::RATE);
-    printf(".");
+    //monitoringCollector->send({ bytesTotal, "readout.BytesTotal" }, DerivedMetricMode::RATE);
+    printf("%d\n",n);
     fflush(stdout);
-    sleep(1);
+    //sleep(1);
   }
 
   return 0;